+++ packages/plugin-file-upload/web/src/toolbar/inline-buttons.tsx
@@ -1,11 +1,8 @@
-import { BUTTONS, PluginSettingsIcon } from 'wix-rich-content-plugin-commons';
+import { BUTTONS, PluginSettingsIcon, Uploader } from 'wix-rich-content-plugin-commons';
 import { MediaReplaceIcon } from '../icons';
 import type {
   CreateInlineButtons,
   TranslationFunction,
-  AnchorTarget,
-  RelValue,
-  UISettings,
   AvailableExperiments,
 } from 'wix-rich-content-common';
 import type { FilePluginEditorConfig } from '../types';
@@ -13,6 +10,7 @@ import { get } from 'lodash';
 import { getModalStyles } from 'wix-rich-content-editor-common';
 import { FILE_UPLOAD_TYPE } from '../types';
 import { Modals } from '../modals';
+import { FilePluginService } from './filePluginService';
 
 const createInlineButtons: CreateInlineButtons = ({
   t,
@@ -46,6 +44,20 @@ const createInlineButtons: CreateInlineButtons = ({
       ]
     : [];
 
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const replaceButton: any = {
+    keyName: 'replace',
+    type: BUTTONS.FILES,
+    icon: icons?.replace || MediaReplaceIcon,
+    settings,
+    tooltipTextKey: t('FileUploadReplaceButton_tooltip'),
+  };
+
+  if (experiments?.useNewUploadContext?.enabled) {
+    replaceButton.mediaPluginService = new FilePluginService();
+    replaceButton.getUploader = () => new Uploader(settings.onFileSelected);
+  }
+
   return [
     { keyName: 'sizeSmall', type: BUTTONS.SIZE_SMALL_CENTER, mobile: false },
     { keyName: 'sizeContent', type: BUTTONS.SIZE_CONTENT, mobile: false },
@@ -54,13 +66,7 @@ const createInlineButtons: CreateInlineButtons = ({
     { keyName: 'alignCenter', type: BUTTONS.ALIGN_CENTER, mobile: false },
     { keyName: 'sizeSmallRight', type: BUTTONS.SIZE_SMALL_RIGHT, mobile: false },
     { keyName: 'separator3', type: BUTTONS.SEPARATOR, mobile: false },
-    {
-      keyName: 'replace',
-      type: BUTTONS.FILES,
-      icon: icons?.replace || MediaReplaceIcon,
-      settings,
-      tooltipTextKey: t('FileUploadReplaceButton_tooltip'),
-    },
+    replaceButton,
     ...settingsButton,
     { keyName: 'delete', type: BUTTONS.DELETE, mobile: true },
   ];